refactor(internal): group 11 periphery packages under internal/platform/ (#341)
ARCHITECTURE-v0.7.10.md §3 names the periphery as "à regrouper plus tard
sous platform/". This is that regroup. No logic change, no behavior
change — purely a mechanical move + import rewrite.

Moved under internal/platform/:
- eventlog, gittrack, media, mood, session, signal,
  supervisor, tlsgen, trace, updater, vulncheck

What did NOT move:
- scheduler/   — the doc tags it as "consumer du Runtime", not periphery
- telegram/    — user-facing consumer, not ops plumbing
- voice/       — user-facing consumer

Rationale: with the 5 blocks (capability/memory/ai/sandbox/runtime) at
the top of internal/ and the consumers clearly named (cli, controlcenter,
telegram, voice, scheduler), having 11 ops packages flat at the same
level made the block structure hard to see. platform/ makes the intent
from §3 visible in the tree.

Mechanics:
- 11 `git mv` for the package directories.
- 20 files' import paths rewritten
  (sed over the 17 known consumer files + 3 cmd/ mains).
- technical/TEST-BASELINE.md paths updated.

Validation:
- go build ./... clean
- TestFoundationDependencyRules / TestConsumerDependencyRules /
  TestHardcodedModelFallback — PASS (archtest doesn't reference these
  packages by path, so no rule changes needed).
- make regression green, global coverage 48.2% (identical to pre-move).
