+++ cmd/alf-daemon/main.go
@@ -25,26 +25,26 @@ import (
 	"github.com/alamparelli/alf/internal/envsecrets"
 	vault "github.com/alamparelli/alf/internal/sandbox/secrets"
 	cc "github.com/alamparelli/alf/internal/controlcenter"
-	"github.com/alamparelli/alf/internal/eventlog"
-	"github.com/alamparelli/alf/internal/gittrack"
-	"github.com/alamparelli/alf/internal/media"
+	"github.com/alamparelli/alf/internal/platform/eventlog"
+	"github.com/alamparelli/alf/internal/platform/gittrack"
+	"github.com/alamparelli/alf/internal/platform/media"
 	"github.com/alamparelli/alf/internal/memory"
 	"github.com/alamparelli/alf/internal/memory/socketsrv"
 	"github.com/alamparelli/alf/internal/memory/curation"
-	"github.com/alamparelli/alf/internal/mood"
+	"github.com/alamparelli/alf/internal/platform/mood"
 	provider "github.com/alamparelli/alf/internal/ai/provider"
 	"github.com/alamparelli/alf/internal/runtime"
 	"github.com/alamparelli/alf/internal/sandbox"
 	"github.com/alamparelli/alf/internal/sandbox/integrity"
 	"github.com/alamparelli/alf/internal/scheduler"
-	"github.com/alamparelli/alf/internal/session"
-	"github.com/alamparelli/alf/internal/signal"
-	"github.com/alamparelli/alf/internal/supervisor"
+	"github.com/alamparelli/alf/internal/platform/session"
+	"github.com/alamparelli/alf/internal/platform/signal"
+	"github.com/alamparelli/alf/internal/platform/supervisor"
 	"github.com/alamparelli/alf/internal/skills"
 	"github.com/alamparelli/alf/internal/tooling"
-	"github.com/alamparelli/alf/internal/trace"
+	"github.com/alamparelli/alf/internal/platform/trace"
 	tgclient "github.com/alamparelli/alf/internal/telegram"
-	"github.com/alamparelli/alf/internal/updater"
+	"github.com/alamparelli/alf/internal/platform/updater"
 	"github.com/alamparelli/alf/internal/voice"
 	"gopkg.in/natefinch/lumberjack.v2"
 )
