+++ internal/controlcenter/chat_service.go
@@ -14,13 +14,13 @@ import (
 	"github.com/alamparelli/alf/internal/ai"
 	provider "github.com/alamparelli/alf/internal/ai/provider"
 	"github.com/alamparelli/alf/internal/comms"
-	"github.com/alamparelli/alf/internal/eventlog"
-	"github.com/alamparelli/alf/internal/media"
+	"github.com/alamparelli/alf/internal/platform/eventlog"
+	"github.com/alamparelli/alf/internal/platform/media"
 	"github.com/alamparelli/alf/internal/memory"
-	"github.com/alamparelli/alf/internal/mood"
+	"github.com/alamparelli/alf/internal/platform/mood"
 	"github.com/alamparelli/alf/internal/runtime"
 	agents "github.com/alamparelli/alf/internal/runtime/agents"
-	chatsession "github.com/alamparelli/alf/internal/session"
+	chatsession "github.com/alamparelli/alf/internal/platform/session"
 	"github.com/alamparelli/alf/internal/skills"
 	"github.com/alamparelli/alf/internal/tooling"
 	"github.com/alamparelli/alf/internal/voice"
