+++ internal/platform/eventlog/logger.go
