+++ internal/platform/eventlog/logger_test.go
