+++ internal/platform/media/detector.go
