+++ internal/platform/media/detector_more_test.go
