+++ internal/platform/media/detector_test.go
