+++ internal/platform/media/frames_integration_test.go
