+++ internal/platform/mood/mood.go
