+++ internal/platform/mood/mood_test.go
