+++ internal/platform/mood/mood_updater.go
