+++ internal/platform/mood/reaction_score.go
