+++ internal/platform/mood/reactions.go
