+++ internal/platform/session/store.go
