+++ internal/platform/session/store_test.go
