+++ internal/platform/signal/server.go
