+++ internal/platform/signal/server_test.go
