+++ internal/platform/supervisor/supervisor_test.go
