+++ internal/platform/trace/stats.go
