+++ internal/platform/trace/stats_test.go
