+++ internal/platform/updater/checker.go
