+++ internal/platform/updater/checker_test.go
