+++ internal/platform/vulncheck/vulncheck_test.go
